% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{adjustIntervals}
\alias{adjustIntervals}
\title{Adjust Date Intervals}
\usage{
adjustIntervals(stationID_proc, procID_proc, intervals,
  fname_df_project = "/gws/nopw/j04/eddystore/eddystore_projects/df_eddystore_projects.csv")
}
\arguments{
\item{stationID_proc}{A character string identifying a station in the project data frame.}

\item{procID_proc}{A character string identifying a processing setup in the project data frame.}

\item{intervals}{A list of the start and end times of each interval, created by makeDateIntervals.}

\item{fname_df_project}{Path to the eddystore project file table.}
}
\value{
An eddystore job object.
}
\description{
This function adjusts the output of makeDateIntervals to match 
boundaries between .eddypro project files.
}
\examples{
stationID_proc <- "EasterBush"
procID_proc <- "CO2_H2O"
nIntervals <- 4
startDate_period <- "2006-07-01 00:00"
endDate_period   <- "2007-12-31 23:30"
startDate_period <- as.POSIXct(strptime(startDate_period, "\%Y-\%m-\%d \%H:\%M"), tz = "UTC")
endDate_period   <- as.POSIXct(strptime(endDate_period,   "\%Y-\%m-\%d \%H:\%M"), tz = "UTC")
fname_df_project = "C:/Users/plevy/Documents/eddystore_projects/df_eddystore_projects.csv"
intervals <- makeDateIntervals(startDate_period, endDate_period, nIntervals)
myJob <- adjustIntervals(stationID_proc, procID_proc, intervals, fname_df_project = fname_df_project)
}
\seealso{
\code{\link{adjustIntervals}} for the adjusting this to match boundaries between processing files.
}
