% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{checkJobCompleted}
\alias{checkJobCompleted}
\title{Check Status of Job on LOTUS - Check if Completed}
\usage{
checkJobCompleted(job_name)
}
\arguments{
\item{job_name}{The job name identifier given in df_job_requests.}
}
\value{
job_status Logical, TRUE if job_status == "DONE"
}
\description{
This function checks an eddypro processing job on LOTUS.
For multi-processor jobs, it only checks the first-listed in the job array, 
may report "DONE" before all are finished.
}
