% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{checkJobRunning}
\alias{checkJobRunning}
\title{Check Status of Job on LOTUS - Check if Still Running}
\usage{
checkJobRunning(job_name)
}
\arguments{
\item{job_name}{An eddystore job_name specified by the job request}
}
\value{
job_status Logical, TRUE if job_status == "RUN" i.e. still running
}
\description{
This function checks an eddypro processing job on LOTUS.
For multi-processor jobs, it only checks the first-listed in the job array,
}
