% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{checkJobStatus}
\alias{checkJobStatus}
\title{Check Status of Job on LOTUS}
\usage{
checkJobStatus(job)
}
\arguments{
\item{job}{An eddystore job object made with the createJob function}
}
\value{
job_status A character message: "RUN" if still running, "DONE" if finished.
}
\description{
This function checks an eddypro processing job on LOTUS.
For multi-processor jobs, it only checks the first-listed in the job array, 
may report "DONE" before all are finished.
}
