% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{convertProjectPath}
\alias{convertProjectPath}
\title{Convert paths in Eddypro project files}
\usage{
convertProjectPath(eddyproProjectPathName, station_dir)
}
\arguments{
\item{eddyproProjectPathName}{Path to the location of the uploaded project file on JASMIN.}

\item{station_dir}{Path to the location of the station directory on JASMIN e.g. "/gws/nopw/j04/eddystore/stations/EasterBush".}
}
\value{
eddyproProjectPathName_new File name for project file with eddystore paths. The side effect is to write this file.
}
\description{
Need to add biomet file path changes biom_file and biom_dir
}
\details{
This function changes all references to paths in an uploaded project file to the eddystore path names.
}
\examples{
eddyproProjectPathName <- "C:/Users/plevy/Documents/eddystore_projects/stations/EasterBush/projects/processing2.eddypro"
station_dir <- "C:/Users/plevy/Documents/eddystore_projects/stations/EasterBush"
convertProjectPath(eddyproProjectPathName, station_dir)
}
