% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{createJob}
\alias{createJob}
\title{Create Processing Job}
\usage{
createJob(stationID_proc, procID_proc, startDate_period, endDate_period,
  nProcessors,
  fname_df_project = "/gws/nopw/j04/eddystore/eddystore_projects/df_eddystore_projects.csv",
  binpath = "/gws/nopw/j04/eddystore/eddypro-engine_6.2.0/eddypro-engine/bin/linux/eddypro_rp",
  switch_OS = "-s linux", eddystore_path = "/gws/nopw/j04/eddystore",
  job_name = "eddytest", user_email = "plevy@ceh.ac.uk")
}
\arguments{
\item{stationID_proc}{A character string identifying a station in the project data frame.}

\item{procID_proc}{A character string identifying a processing setup in the project data frame.}

\item{startDate_period}{A character string for the time the run was started, used to identify the relevant files.}

\item{endDate_period}{A character string for the time the run was started, used to identify the relevant files.}

\item{nProcessors}{An integer number of processors to use, equal to the number of time intervals to split the processing into.}

\item{fname_df_project}{Path to the eddystore project file table.}

\item{binpath}{Path to the eddy_rp raw data processing binary file.}

\item{switch_OS}{Switch betweeen linux and windows versions.}

\item{eddystore_path}{The path to eddystore.}

\item{job_name}{The job name identifier given in df_job_requests.}

\item{user_email}{E-mail address to send notifications to.}
}
\value{
An eddystore job object.
}
\description{
This function creates an eddypro processing job to be run on LOTUS.
}
\examples{
stationID_proc <- "EasterBush"
procID_proc <- "CO2_H2O"
nProcessors <- 4
startDate_period <- "2006-07-01 00:00"
endDate_period   <- "2007-12-31 23:30"
startDate_period <- as.POSIXct(strptime(startDate_period, "\%Y-\%m-\%d \%H:\%M"), tz = "UTC")
endDate_period   <- as.POSIXct(strptime(endDate_period,   "\%Y-\%m-\%d \%H:\%M"), tz = "UTC")
myJob <- createJob(stationID_proc, procID_proc, startDate_period, endDate_period, nProcessors,
    fname_df_project = "C:/Users/plevy/Documents/eddystore_projects/df_eddystore_projects.csv", 
    eddystore_path = "N:/0Peter/curr/ECsystem/eddystore",
    job_name = "createJob_example", user_email = "plevy@ceh.ac.uk")
}
\seealso{
\code{\link{adjustIntervals}} for the adjusting this to match boundaries between project files.
}
