% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\docType{package}
\name{eddystore-package}
\alias{eddystore}
\alias{eddystore-package}
\title{Parallel processing of eddy covariance data on JASMIN.}
\description{
eddystore allows you to set up and run processing jobs
to calculate eddy covariance fluxes using eddypro on JASMIN.
}
\author{
\strong{Maintainer}: Peter Levy \email{plevy@ceh.ac.uk}

}
