% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{get_essential_output_df}
\alias{get_essential_output_df}
\title{Collate essential output files}
\usage{
get_essential_output_df(job_time)
}
\arguments{
\item{job_time}{A character string for the time the run was started, used to identify the relevant files.}
}
\description{
Note that curently "EasterBush" is hard-coded in the path - need to pass path argument
This function collates all the essential output files.
}
