% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{get_essential_output_df_byTime}
\alias{get_essential_output_df_byTime}
\title{Collate essential output files using time only}
\usage{
get_essential_output_df_byTime(job_startTime)
}
\arguments{
\item{job_startTime}{Start time of the eddystore job made with the runJob function}
}
\value{
A data frame of the concatenated essential output files from each job.
}
\description{
Note that curently "EasterBush" is hard-coded in the path - need to pass path argument
}
