% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{makeDateIntervals}
\alias{makeDateIntervals}
\title{Make Date Intervals}
\usage{
makeDateIntervals(startDate_period, endDate_period, nIntervals)
}
\arguments{
\item{startDate_period}{A character string for the time the run was started, used to identify the relevant files.}

\item{endDate_period}{A character string for the time the run was started, used to identify the relevant files.}

\item{nIntervals}{An integer number of intervals to use to split the processing into.}
}
\description{
This function creates a number of equally sized time intervals between 
the specified start and end times.
}
\examples{
nIntervals = 4
startDate_proc <- "2007-01-01 00:00"
endDate_proc   <- "2007-12-31 23:30"
startDate_proc <- as.POSIXct(strptime(startDate_proc, "\%Y-\%m-\%d \%H:\%M"), tz = "UTC")
endDate_proc   <- as.POSIXct(strptime(endDate_proc,   "\%Y-\%m-\%d \%H:\%M"), tz = "UTC")
intervals <- makeDateIntervals(startDate_proc, endDate_proc, nIntervals)
}
\seealso{
\code{\link{adjustIntervals}} for the adjusting this to match boundaries between processing files.
}
