% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{runJob}
\alias{runJob}
\title{Run Processing Job}
\usage{
runJob(jobFileName)
}
\arguments{
\item{job}{An eddystore job object made with the createJob function}
}
\description{
This function runs an eddypro processing job on LOTUS.
Gives: Error in system(command, as.integer(flag), f, stdout, stderr, timeout) : 
  character string expected as first argument
}
\examples{
myJob <- runJob(myJob)
}
\seealso{
\code{\link{adjustIntervals}} for the adjusting this to match boundaries between project files.
}
