% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{writeEddyproProcFilesForIntervals}
\alias{writeEddyproProcFilesForIntervals}
\title{Write Eddypro Processing files for all intervals}
\usage{
writeEddyproProcFilesForIntervals(eddyproProcFileName, intervals)
}
\arguments{
\item{eddyproProcFileName}{A character string identifying the root eddypro file.}

\item{intervals}{A list of the start and end times of each interval, created by makeDateIntervals.}
}
\description{
This function writes an eddypro processing file for each of the intervals specified.
Should eddyproProcFileName actually be looked up in the proc table by stationID_proc, procID_proc in adjustIntervals, 
and added to the object returned there?
}
\examples{
eddyproProcFileName <- "/group_workspaces/jasmin2/eddystore/stations/EasterBush/proc/processing.eddypro"
v_EddyproProcFileNames <- writeEddyproProcFilesForIntervals(eddyproProcFileName, intervals)
}
\seealso{
\code{\link{adjustIntervals}} for the adjusting this to match boundaries between processing files.
}
