% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{writeProjectFiles}
\alias{writeProjectFiles}
\title{Write Eddypro Project files for all intervals}
\usage{
writeProjectFiles(job)
}
\arguments{
\item{job}{A list of the start and end times of each interval, created by makeDateIntervals.}
}
\description{
This function writes an eddypro project file for each of the intervals specified.
}
\examples{
eddyproProcFileName <- "/group_workspaces/jasmin2/eddystore/stations/EasterBush/proc/processing.eddypro"
myJob <- writeProjectFiles(myJob)
}
\seealso{
\code{\link{adjustIntervals}} for the adjusting this to match boundaries between processing files.
}
