% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{adjustIntervals}
\alias{adjustIntervals}
\title{Adjust Date Intervals}
\usage{
adjustIntervals(stationID_proc, procID_proc, intervals)
}
\arguments{
\item{stationID_proc}{A character string identifying a station in the project data frame.}

\item{procID_proc}{A character string identifying a processing setup in the project data frame.}

\item{intervals}{A list of the start and end times of each interval, created by makeDateIntervals.}
}
\description{
This function adjusts the output of makeDateIntervals to match 
boundaries between .eddypro project files.
}
\examples{
myJob <- adjustIntervals(stationID_proc, procID_proc, intervals)
}
\seealso{
\code{\link{adjustIntervals}} for the adjusting this to match boundaries between processing files.
}
