% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{convertProjectPath}
\alias{convertProjectPath}
\title{Convert paths in Eddypro project files}
\usage{
convertProjectPath(eddyproProjectPathName, station_dir)
}
\arguments{
\item{station_dir}{A character string for the path to the location of the station directory on JASMIN.}

\item{filename}{A character string for the path to the location of the station directory on JASMIN.}
}
\value{
eddyproProjectPathName_new File name for project file with eddystore paths. The side effect is to write this file.
}
\description{
Need to add biomet file path changes biom_file and biom_dir
}
\details{
This function changes all references to paths in uploaded project files to the eddystore path names.
}
\examples{
eddyproProjectPathName <- "C:/Users/plevy/Documents/eddystore_procdata/stations/EasterBush/projects/processing2.eddypro"
station_dir <- "C:/Users/plevy/Documents/eddystore_procdata/stations/EasterBush"
convertProjectPath(eddyproProjectPathName, station_dir)
}
