% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{createJob}
\alias{createJob}
\title{Create Processing Job}
\usage{
createJob(stationID_proc, procID_proc, startDate_period, endDate_period,
  nProcessors)
}
\arguments{
\item{stationID_proc}{A character string identifying a station in the project data frame.}

\item{procID_proc}{A character string identifying a processing setup in the project data frame.}

\item{startDate_period}{A character string for the time the run was started, used to identify the relevant files.}

\item{endDate_period}{A character string for the time the run was started, used to identify the relevant files.}

\item{nProcessors}{An integer number of processors to use, equal to the number of time intervals to split the processing into.}
}
\description{
This function creates an eddypro processing job to be run on LOTUS.
}
\examples{
myJob <- createJob(stationID_proc, procID_proc, startDate_period, endDate_period, nProcessors)
}
\seealso{
\code{\link{adjustIntervals}} for the adjusting this to match boundaries between project files.
}
