% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{createJob}
\alias{createJob}
\title{Create Processing Job}
\usage{
createJob(stationID_proc, procID_proc, startDate_period, endDate_period,
  nProcessors)
}
\arguments{
\item{stationID_proc}{A character string identifying a station in the project data frame.}

\item{procID_proc}{A character string identifying a processing setup in the project data frame.}

\item{startDate_period}{A character string for the time the run was started, used to identify the relevant files.}

\item{endDate_period}{A character string for the time the run was started, used to identify the relevant files.}

\item{nProcessors}{An integer number of processors to use, equal to the number of time intervals to split the processing into.}
}
\description{
This function creates an eddypro processing job to be run on LOTUS.
}
\examples{
stationID_proc <- "EasterBush"
procID_proc <- "CO2_H2O"
nProcessors <- 4
startDate_period <- "2006-07-01 00:00"
endDate_period   <- "2007-12-31 23:30"
startDate_period <- as.POSIXct(strptime(startDate_period, "\%Y-\%m-\%d \%H:\%M"), tz = "UTC")
endDate_period   <- as.POSIXct(strptime(endDate_period,   "\%Y-\%m-\%d \%H:\%M"), tz = "UTC")
# read eddypro project data on local disk
df_project <- read.csv("C:/Users/plevy/Documents/eddystore_procdata/eddystore_proc_table.csv", stringsAsFactors = FALSE)
df_project$startDate <- as.POSIXct(strptime(df_project$startDate, "\%d/\%m/\%Y \%H:\%M"), tz = "UTC")
df_project$endDate   <- as.POSIXct(strptime(df_project$endDate, "\%d/\%m/\%Y \%H:\%M"), tz = "UTC")
myJob <- createJob(stationID_proc, procID_proc, startDate_period, endDate_period, nProcessors)
}
\seealso{
\code{\link{adjustIntervals}} for the adjusting this to match boundaries between project files.
}
