% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{get_essential_output_df}
\alias{get_essential_output_df}
\title{Collate essential output files using job name}
\usage{
get_essential_output_df(job_name, station_dir)
}
\arguments{
\item{job_name}{The job name identifier given in df_job_requests.}

\item{station_dir}{The path to the station directory on eddystore.}
}
\value{
A data frame of the concatenated essential output files from each job.
}
\description{
Collate essential output files using job name
}
