% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{get_essential_output_df}
\alias{get_essential_output_df}
\title{Collate essential output files}
\usage{
get_essential_output_df(job)
}
\arguments{
\item{job}{An eddystore job object made with the createJob function}
}
\value{
A data frame of the concatenated essential output files from each job.
}
\description{
Note that curently "EasterBush" is hard-coded in the path - need to pass path argument
}
