% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{runJob}
\alias{runJob}
\title{Run Processing Job}
\usage{
runJob(job)
}
\arguments{
\item{job}{An eddystore job object made with the createJob function}
}
\description{
This function runs an eddypro processing job on LOTUS.
Gives: Error in system(command, as.integer(flag), f, stdout, stderr, timeout) : 
  character string expected as first argument
}
\examples{
stationID_proc <- "EasterBush"
procID_proc <- "CO2_H2O"
nProcessors <- 4
startDate_period <- "2006-07-01 00:00"
endDate_period   <- "2007-12-31 23:30"
startDate_period <- as.POSIXct(strptime(startDate_period, "\%Y-\%m-\%d \%H:\%M"), tz = "UTC")
endDate_period   <- as.POSIXct(strptime(endDate_period,   "\%Y-\%m-\%d \%H:\%M"), tz = "UTC")
df_project <- read_excel("C:/Users/plevy/Documents/eddystore/data-raw/eddystore_proc_table.xlsx")
myJob <- createJob(stationID_proc, procID_proc, startDate_period, endDate_period, nProcessors)
myJob <- runJob(myJob)
}
\seealso{
\code{\link{adjustIntervals}} for the adjusting this to match boundaries between project files.
}
