% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{writeJobFile}
\alias{writeJobFile}
\title{Write LOTUS batch job files for all intervals}
\usage{
writeJobFile(job)
}
\arguments{
\item{eddyproProcFileName}{A character string identifying the root eddypro file.}

\item{nIntervals}{An integer number of intervals to use to split the processing into.}
}
\description{
This function writes a LOTUS batch job file for each of the intervals specified.
Should eddyproProcFileName actually be looked up in the proc table by stationID_proc, procID_proc in adjustIntervals, 
and added to the object returned there?
}
\examples{
jobFileName <- writeJobFile(myJob)
jobFileName
}
\seealso{
\code{\link{adjustIntervals}} for the adjusting this to match boundaries between processing files.
}
