% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{writeJobFile}
\alias{writeJobFile}
\title{Write LOTUS batch job files for all intervals}
\usage{
writeJobFile(job,
  binpath = "/gws/nopw/j04/eddystore/eddypro-engine_6.2.0/eddypro-engine/bin/linux/eddypro_rp",
  switch_OS = "-s linux", eddystore_path = "/gws/nopw/j04/eddystore",
  user_email = "plevy@ceh.ac.uk")
}
\arguments{
\item{job}{An eddystore job object made with the writeProjectFiles function.}

\item{binpath}{Path to the eddy_rp raw data processing binary file.}

\item{switch_OS}{Switch betweeen linux and windows versions.}

\item{eddystore_path}{The path to eddystore.}

\item{user_email}{E-mail address to send notifications to.}
}
\value{
An eddystore job object.
}
\description{
This function writes a LOTUS batch job file for each of the intervals specified.
}
\examples{
stationID_proc <- "EasterBush"
procID_proc <- "CO2_H2O"
nIntervals <- 4
startDate_period <- "2006-07-01 00:00"
endDate_period   <- "2007-12-31 23:30"
startDate_period <- as.POSIXct(strptime(startDate_period, "\%Y-\%m-\%d \%H:\%M"), tz = "UTC")
endDate_period   <- as.POSIXct(strptime(endDate_period,   "\%Y-\%m-\%d \%H:\%M"), tz = "UTC")
intervals <- makeDateIntervals(startDate_period, endDate_period, nIntervals)
fname_df_project = "C:/Users/plevy/Documents/eddystore_projects/df_eddystore_projects.csv"
myJob <- adjustIntervals(stationID_proc, procID_proc, intervals, fname_df_project = fname_df_project)
myJob <- writeProjectFiles(myJob)
myJob <- writeJobFile(myJob, binpath = "N:/0Peter/curr/ECsystem/eddypro", 
                             switch_OS = "-s linux",
                             eddystore_path = "N:/0Peter/curr/ECsystem/eddystore")
}
\seealso{
\code{\link{adjustIntervals}} for the adjusting this to match boundaries between processing files.
}
