% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddystore.R
\name{writeProjectFiles}
\alias{writeProjectFiles}
\title{Write Eddypro Project files for all intervals}
\usage{
writeProjectFiles(job)
}
\arguments{
\item{job}{A list of the start and end times of each interval, created by makeDateIntervals.}
}
\value{
An eddystore job object.
}
\description{
This function writes an eddypro project file for each of the intervals specified.
}
\examples{
stationID_proc <- "EasterBush"
procID_proc <- "CO2_H2O"
nIntervals <- 4
startDate_period <- "2006-07-01 00:00"
endDate_period   <- "2007-12-31 23:30"
startDate_period <- as.POSIXct(strptime(startDate_period, "\%Y-\%m-\%d \%H:\%M"), tz = "UTC")
endDate_period   <- as.POSIXct(strptime(endDate_period,   "\%Y-\%m-\%d \%H:\%M"), tz = "UTC")
intervals <- makeDateIntervals(startDate_period, endDate_period, nIntervals)
fname_df_project = "C:/Users/plevy/Documents/eddystore_projects/df_eddystore_projects.csv"
myJob <- adjustIntervals(stationID_proc, procID_proc, intervals, fname_df_project = fname_df_project)
myJob <- writeProjectFiles(myJob)
}
\seealso{
\code{\link{adjustIntervals}} for the adjusting this to match boundaries between processing files.
}
